+++ plugins/pm_filter.py
@@ -30,8 +30,6 @@ from database.filters_mdb import (
     find_filter,
     get_filters,
 )
-import openai
-import os
 import logging
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.ERROR)
@@ -39,37 +37,13 @@ logger.setLevel(logging.ERROR)
 req_channel = REQ_CHANNEL
 BUTTONS = {}
 SPELL_CHECK = {}
-openai.api_key = OPENAI_API
 
-@Client.on_message(filters.private & filters.text & filters.incoming)
-async def lazy_answer(client, message):
-    lazy_users_message = message.text
-    user_id = message.from_user.id
-    response = openai.Completion.create(
-        model = "text-davinci-003",
-        prompt = lazy_users_message,
-        temperature = 0.5, 
-        max_tokens = 1000,
-        top_p=1,
-        frequency_penalty=0.1,
-        presence_penalty = 0.0,
-    )
-    btn=[
-            [InlineKeyboardButton(text=f"⇱🤷‍♀️ Take Action 🗃️⇲", url=f'https://t.me/{temp.U_NAME}')],
-            [InlineKeyboardButton(text=f"🗑 Delete log ❌", callback_data=f'close_data')],
-         ]
-    reply_markup=InlineKeyboardMarkup(btn)
-    footer_credit = "❚█══<a href='https://telegram.me/LazyDeveloperSupport'>𝘙𝘌𝘗𝘖𝘙𝘛 𝘐𝘚𝘚𝘜𝘌</a>═══════\n❚█══<a href='https://telegram.me/LazyDeveloperr'>𝘊𝘖𝘕𝘛𝘈𝘊𝘛 𝘔𝘈𝘚𝘛𝘌𝘙</a>═══════"
-    lazy_response = response.choices[0].text 
-    await client.send_message(LAZY_AI_LOGS, text=f"⚡️#Lazy_AI_Query \n\n• A user named **{message.from_user.mention}** with user id - `{user_id}`. Asked me this query...\n\══❚█══Q࿐U࿐E࿐R࿐Y══█❚══\n\n[Q྿.]**{lazy_users_message}**\n\n◔̯◔Here is what i responded:\n[A྿.] `{lazy_response}`\n\n\n █❚═USER ID═❚═• `{user_id}` \n█❚═USER Name═❚═• `{message.from_user.mention}` \n\n🗃️" , reply_markup = reply_markup )
-    await message.reply(f"{lazy_response}\n\n\n{footer_credit}")
 
 @Client.on_message(filters.group & filters.text & filters.incoming)
 async def give_filter(client, message):
     k = await manual_filters(client, message)
     if k == False:
         await auto_filter(client, message)
-
 
 @Client.on_callback_query(filters.regex('rename'))
 async def rename(bot,update):
